[EPO-793] First attempt at auto-running cells.
If you mark a cell as autoRun: true, now it will run on loading of
the notebook.

I've taken out the hook to rerun on change, because if you switch
between two notebooks then the code will be re-run (eliminating
answers put into answer boxes, and selections on graphs).

I've also elimited the else, because sometimes we want to hide the
code, have the code be read only, and have the code run automatically.